Adds extra config, specifically for containers
