+++ src/signals/incident/definitions/wizard-step-2-vulaan/afval-container.test.ts
@@ -1,0 +1,9 @@
+import afvalContainer from './afval-container'
+
+describe('definition afvalContainer', () => {
+  it('has a defined set of controls', () => {
+    const keys = Object.keys(afvalContainer)
+
+    expect(keys).toStrictEqual(['extra_container'])
+  })
+})
