+++ src/signals/incident/definitions/wizard-step-2-vulaan/afval.test.ts
@@ -4,6 +4,6 @@ describe('definition afval', () => {
   it('has a defined set of controls', () => {
     const keys = Object.keys(afval)
 
-    expect(keys).toStrictEqual(['locatie', 'extra_afval', 'extra_container'])
+    expect(keys).toStrictEqual(['locatie', 'extra_afval'])
   })
 })
