+++ src/signals/incident/definitions/wizard-step-2-vulaan/civieleConstructies.test.ts
@@ -1,6 +1,6 @@
 import civieleConstructies from './civieleConstructies'
 
-describe('definition civieleConstructiesn', () => {
+describe('definition civieleConstructies', () => {
   it('has a defined set of controls', () => {
     const keys = Object.keys(civieleConstructies)
